project Recipes, accomplishment stage 5 v3

+++ Recipes/Recipes/task/src/recipes/servicies/RecipeService.java
@@ -1,9 +1,15 @@
-package recipes;
+package recipes.servicies;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
+import recipes.entites.Recipe;
+import recipes.repositories.RecipeRepository;
+import recipes.repositories.UserRepository;
+import recipes.securityconfig.IAuthenticationFacade;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
@@ -12,10 +18,15 @@ import java.util.List;
 @Service
 public class RecipeService {
 
-    private final RecipeRepository recipeRepository;
+    @Autowired
+    private IAuthenticationFacade authenticationFacade;
 
-    public RecipeService(RecipeRepository recipeRepository) {
+    private final RecipeRepository recipeRepository;
+    private final UserRepository userRepository;
+
+    public RecipeService(RecipeRepository recipeRepository, UserRepository userRepository) {
         this.recipeRepository = recipeRepository;
+        this.userRepository = userRepository;
     }
 
     public Recipe findRecipeById(Long id) {
@@ -24,6 +35,8 @@ public class RecipeService {
     }
 
     public String addNewRecipe(Recipe recipe) {
+        Authentication authentication = authenticationFacade.getAuthentication();
+        recipe.setAuthor(authentication.getName());
         recipe.setDate(LocalDateTime.now());
         recipeRepository.save(recipe);
         return "{\"id\": " + recipe.getId() + "}";
@@ -31,14 +44,19 @@ public class RecipeService {
 
     public ResponseEntity deleteRecipeById(Long id) {
         if (recipeRepository.findById(id).isPresent()) {
-            recipeRepository.deleteById(id);
+            Authentication authentication = authenticationFacade.getAuthentication();
+            if (recipeRepository.findById(id).get().getAuthor().equals(authentication.getName())) {
+                recipeRepository.deleteById(id);
+            } else return new ResponseEntity<>(HttpStatus.FORBIDDEN);
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
         } else return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
     public ResponseEntity updateRecipeById(Long id, Recipe recipe) {
         Recipe recipeFromDb = recipeRepository.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
+        Authentication authentication = authenticationFacade.getAuthentication();
         recipeFromDb = recipe;
+        recipeFromDb.setAuthor(authentication.getName());
         recipeFromDb.setId(id);
         recipeFromDb.setDate(LocalDateTime.now());
         recipeRepository.save(recipeFromDb);
